refactor: rename rp
