+++ src/components/RedPacket/RedPacketCardFooter.tsx
