+++ src/mask/plugins/red-packet/components/RedPacketMessagePanel.tsx
