+++ src/mask/plugins/red-packet/hooks/useCreateFTRedPacketCallback.ts
