+++ src/modals/RedPacketModal/ConfirmView.tsx
