+++ src/modals/RedPacketModal/HistoryDetailView.tsx
