+++ src/modals/RedPacketModal/MainView.tsx
