+++ src/modals/RedPacketModal/RequirementsView.tsx
