+++ src/modals/RedPacketModal/RootView.tsx
