+++ src/providers/red-packet/types.ts
@@ -227,7 +227,7 @@ export namespace FireflyRedPacketAPI {
         tid: string;
         cover: ThemeSettings;
         normal: ThemeSettings;
-        /** Redpacket without theme settings preset, current ones are default */
+        /** RedPacket without theme settings preset, current ones are default */
         is_default?: boolean;
     }
